+++ src/fastauth/__init__.py
@@ -70,4 +70,4 @@ __all__ = [
     "openapi",
     "test_utils",
 ]
-__version__ = "0.10.0"
+__version__ = "0.10.1"